requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
SQLAlchemy==2.0.31
tldextract==5.1.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
zope.event==5.0
zope.interface==6.4.post2
